Added possibility for better gray-scale mathing in nr stitcher. Now the stitcher is able to match standard deviations of gray values between tiles, in addition to means.

+++ itl2/progress.cpp
@@ -19,7 +19,7 @@ namespace itl2
 					prog.step();
 			}
 
-			float MAX = 2235450000.0f;
+			size_t MAX = 2235450000;
 			{
 				ProgressIndicator prog(MAX);
 				for (float n = 0; n < 3 * MAX / 4; n += 173)
